Module 4 completed
In this module, i have added the function createPhoneIntent in the InfoActivity to handle the phone call with a loaded phone number on the dial screen. I have also add the following properties in the activity_info.xml file (onClick, clickable, focusable) on the text view with id = text_view_phone.